financial/requirements.txt
Flask==2.0.2
helios-opentelemetry-sdk==0.1.67
werkzeug==2.0.3
